requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
crispy==0.7.4
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==40.0.2
cssselect2==0.7.0
cycler==0.11.0
Django==4.2.1
django-crispy-forms==2.0
django-filter==23.2
fabio==2023.4.1
fonttools==4.39.4
h5py==3.8.0
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pyparsing==3.0.9
pypdf==3.8.1
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-bidi==0.4.2
python-dateutil==2.8.2
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.30.0
silx==1.1.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0
uritools==4.0.1
urllib3==2.0.2
webencodings==0.5.1
xhtml2pdf==0.2.11
